norminette, faltou ajustar comentários
