+++ src/garbage/gc_utils.c
@@ -2,41 +2,6 @@
 
 #include "../../include/minishell.h"
 
-// vai para garbage_collector.c
-t_gc	*gc_init(void)
-{
-	t_gc	*gc;
-
-	gc = malloc(sizeof(t_gc));
-	if (!gc)
-		return (NULL);
-	gc->head = NULL;
-	return (gc);
-}
-
-void	gc_add(t_gc *gc, void *ptr)
-{
-	t_gc_node	*new_node;
-
-	new_node = malloc(sizeof(t_gc_node));
-	if (!new_node)
-		return ;
-	new_node->ptr = ptr;
-	new_node->next = gc->head;
-	gc->head = new_node;
-}
-
-void	*gc_malloc(t_gc *gc, size_t size)
-{
-	void	*ptr;
-
-	ptr = malloc(size);
-	if (!ptr)
-		return (NULL);
-	gc_add(gc, ptr);
-	return (ptr);
-}
-
 void	gc_clear(t_gc *gc)
 {
 	t_gc_node	*tmp;
@@ -93,24 +58,6 @@ char	*gc_strdup(const char *s, t_gc *gc)
 	return (dup);
 }
 
-void	*gc_calloc(size_t nmemb, size_t size, t_gc *gc)
-{
-	unsigned char	*temp;
-	size_t			total_size;
-	size_t			i;
-
-	total_size = nmemb * size;
-	if (nmemb != 0 && (total_size / nmemb != size))
-		return (NULL);
-	temp = gc_malloc(gc, (nmemb * size));
-	if (!temp)
-		return (NULL);
-	i = 0;
-	while (i < total_size)
-		temp[i++] = 0;
-	return (temp);
-}
-
 char	*gc_substr(char const *s, unsigned int start, size_t len, t_gc *gc)
 {
 	char		*substr;
@@ -133,4 +80,15 @@ char	*gc_substr(char const *s, unsigned int start, size_t len, t_gc *gc)
 	}
 	substr[i] = '\0';
 	return (substr);
+}
+
+char	*gc_strjoin(char *s1, char *s2, t_gc *gc)
+{
+	char	*joined;
+
+	joined = ft_strjoin(s1, s2);
+	if (!joined)
+		return (NULL);
+	gc_add(gc, joined);
+	return (joined);
 }
