+++ src/validate/validate_utils_2.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   list_utils.c                                       :+:      :+:    :+:   */
+/*   validate_utils_2.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jemorais <jemorais@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jeff <jeff@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/30 15:48:47 by ancarol9          #+#    #+#             */
-/*   Updated: 2025/06/16 16:50:36 by jemorais         ###   ########.fr       */
+/*   Updated: 2025/06/20 17:33:14 by jeff             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,12 @@
 
 #include "../../include/minishell.h"
 
-t_token	*ft_token_last(t_token *lst)
+// tirei de utils mas provavelmente deve voltar
+int	syntax_error(char *msg, t_data *data)
 {
-	t_token	*last;
-
-	if (!lst)
-		return (NULL);
-	last = lst;
-	while (last->next)
-		last = last->next;
-	return (last);
+	ft_putstr_fd("Error: syntaxe error near ", STDERR_FILENO);
+	ft_putstr_fd(msg, STDERR_FILENO);
+	write(STDERR_FILENO, "\n", 1);
+	data->exit_status = 2;
+	return (1);
 }
